feat: replace getPublicConvex with preloadQuery pattern across all pages
- Convert event page to use preloadQuery with EventClient component
- Convert explore page to use preloadQuery with ExploreClient component
- Convert UserInfo to use preloadQuery with UserInfoClient component
- Remove getPublicConvex from convex-server.ts as it's no longer needed
- Update upcoming page to use preloadQuery in generateMetadata

This enables real-time updates across all pages while maintaining SSR benefits.
All pages now automatically update when data changes in Convex without requiring
page refreshes.
