+++ apps/web/app/(base)/event/[eventId]/page.tsx
@@ -1,18 +1,12 @@
 import type { Metadata, ResolvingMetadata } from "next/types";
 import { currentUser } from "@clerk/nextjs/server";
+import { preloadQuery, preloadedQueryResult } from "convex/nextjs";
 
-import type { EventMetadata } from "@soonlist/cal";
-import type {
-  AddToCalendarButtonProps,
-  AddToCalendarButtonPropsRestricted,
-} from "@soonlist/cal/types";
+import type { AddToCalendarButtonProps } from "@soonlist/cal/types";
 import { api } from "@soonlist/backend/convex/_generated/api";
-import { collapseSimilarEvents } from "@soonlist/cal";
 
-import type { EventWithUser } from "~/components/EventList";
-import { EventPage } from "~/components/EventDisplays";
 import { env } from "~/env";
-import { getPublicConvex } from "~/lib/convex-server";
+import { EventClient } from "./EventClient";
 
 interface Props {
   params: Promise<{
@@ -20,33 +14,14 @@ interface Props {
   }>;
 }
 
-// Transform Convex event to EventWithUser format
-function transformConvexEvent(event: any): EventWithUser {
-  return {
-    id: event._id,
-    userId: event.userId,
-    updatedAt: event.updatedAt ? new Date(event.updatedAt) : null,
-    userName: event.userName,
-    event: event.event,
-    eventMetadata: event.eventMetadata,
-    endDateTime: new Date(event.endDateTime),
-    startDateTime: new Date(event.startDateTime),
-    visibility: event.visibility,
-    createdAt: new Date(event._creationTime),
-    user: event.user,
-    eventFollows: event.eventFollows || [],
-    comments: event.comments || [],
-    eventToLists: event.eventToLists || [],
-  };
-}
 
 export async function generateMetadata(
   props: Props,
   parent: ResolvingMetadata,
 ): Promise<Metadata> {
   const params = await props.params;
-  const convex = await getPublicConvex();
-  const event = await convex.query(api.events.get, { eventId: params.eventId });
+  const preloadedEvent = await preloadQuery(api.events.get, { eventId: params.eventId });
+  const event = preloadedQueryResult(preloadedEvent);
   if (!event) {
     return {
       title: "No event found | Soonlist",
@@ -82,70 +57,26 @@ export async function generateMetadata(
 
 export default async function Page(props: Props) {
   const params = await props.params;
-  const convex = await getPublicConvex();
-  const event = await convex.query(api.events.get, { eventId: params.eventId });
   const user = await currentUser();
+
+  const preloadedEvent = await preloadQuery(api.events.get, { eventId: params.eventId });
+  const event = preloadedQueryResult(preloadedEvent);
+  
   if (!event) {
     return <p className="text-lg text-gray-500">No event found.</p>;
   }
-  const eventData = event.event as AddToCalendarButtonPropsRestricted;
-  const eventMetadata = event.eventMetadata as EventMetadata;
-  const fullImageUrl = eventData.images?.[3];
 
-  const convexDuplicates = await convex.query(
+  const preloadedDuplicates = await preloadQuery(
     api.events.getPossibleDuplicates,
-    {
-      startDateTime: event.startDateTime,
-    },
+    { startDateTime: event.startDateTime }
   );
-  const possibleDuplicateEvents = convexDuplicates.map(transformConvexEvent);
-
-  // find the event that matches the current event
-  const similarEvents = collapseSimilarEvents(
-    possibleDuplicateEvents,
-    user?.id,
-  ).find((similarEvent) => similarEvent.event.id === event._id)?.similarEvents;
-
-  // TODO: Implement event lists when list functionality is added to Convex
-  const lists = [];
 
   return (
-    <>
-      <EventPage
-        user={
-          event.user
-            ? {
-                ...event.user,
-                createdAt: new Date(
-                  event.user.created_at || event.user._creationTime,
-                ),
-                updatedAt: event.user.updatedAt
-                  ? new Date(event.user.updatedAt)
-                  : null,
-                onboardingCompletedAt: event.user.onboardingCompletedAt
-                  ? new Date(event.user.onboardingCompletedAt)
-                  : null,
-              }
-            : undefined
-        }
-        eventFollows={event.eventFollows || []}
-        comments={(event.comments || []).map((comment: any) => ({
-          ...comment,
-          createdAt: new Date(comment.created_at || comment._creationTime),
-          updatedAt: comment.updatedAt ? new Date(comment.updatedAt) : null,
-        }))}
-        key={event._id}
-        id={event._id}
-        event={eventData}
-        eventMetadata={eventMetadata}
-        createdAt={new Date(event._creationTime)}
-        visibility={event.visibility}
-        similarEvents={similarEvents}
-        image={fullImageUrl}
-        singleEvent
-        hideCurator
-        lists={[]}
-      />
-    </>
+    <EventClient
+      eventId={params.eventId}
+      userId={user?.id}
+      preloadedEvent={preloadedEvent}
+      preloadedDuplicates={preloadedDuplicates}
+    />
   );
 }
